accounts sytem works now see more ->
+ Changequeue works now.
+ Added edit page that works.
+ Added a seeder for accounts to generate test accounts.
	[!] this will be removed ^
+ Added/Tweakes some buttons on the account listing
+ Added some hyperlinks that doesnt work yet, but they'll soon.
- Removed the Error pages in routes, because it was there for testing.

+++ database/migrations/2018_05_24_074727_create_accounts_table.php
@@ -23,7 +23,7 @@ class CreateAccountsTable extends Migration
             $table->string('ReadGroup');
             $table->string('Tagid');
             $table->longtext('Description');
-            $table->float('ChangeQueue');
+            $table->integer('ChangeQueue');
             $table->timestamps();
             $table->softDeletes();
         });
